requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Faker==6.5.0
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
